requirements.txt
astroid==2.3.3
autopep8==1.5
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.1
python-dotenv==0.11.0
six==1.14.0
typed-ast==1.4.1
Werkzeug==0.16.1
wrapt==1.11.2
